Encourage the use of status codes for unsafe requests (#40)
This PR checks that `render` and `redirect_to` are provided an explicit
status code in response to [unsafe request
methods](https://developer.mozilla.org/en-US/docs/Glossary/Safe/HTTP)
(e.g. POST/PUT/PATCH/DELETE).

#### `redirect_to`

When redirecting a POST/PUT/PATCH/DELETE to a GET location, we should
use 303, rather than a 302. A 302 means that the request method should
not be altered, but:

> Many web browsers implemented this code in a manner that violated this
standard, changing the request type of the new request to
[GET](https://en.wikipedia.org/wiki/HTTP_GET_request), regardless of the
type employed in the original request (e.g.
[POST](https://en.wikipedia.org/wiki/POST_(HTTP))).[[1]](https://en.wikipedia.org/wiki/HTTP_302#cite_note-1)
For this reason, HTTP/1.1 (RFC
[2616](https://datatracker.ietf.org/doc/html/rfc2616)) added the new
status codes [303](https://en.wikipedia.org/wiki/HTTP_303) and
[307](https://en.wikipedia.org/wiki/HTTP_307) to disambiguate between
the two behaviours, with 303 mandating the change of request type to
GET, and 307 preserving the request type as originally sent.
> -- https://en.wikipedia.org/wiki/HTTP_302

Sinatra's [`redirect`
method](https://github.com/sinatra/sinatra/blob/5640495babcb4cfd69ba650b293660b7446402da/lib/sinatra/base.rb#L307-L321)
automatically uses 303 for non-GET requests. [This
PR](https://github.com/rails/rails/pull/45393) aims to do this for Rails
applications.

When autocorrect is enabled, `status: :see_other` will be added to
redirects used in the `create`, `update`, and `delete` actions.

#### `render`

Calling `render` in the `create`, `update`, and `destroy` actions is
almost always associated with error handling, so a 4xx status should be
used. In the vast majority of cases, we should be using 422
Unprocessable Entity.

```ruby
if something.save
  redirect_to something, status: :see_other
else
  render :new, status: :unprocessable_entity
end
```

There are exceptions, though. For example, sometimes you want to show a
confirmation page that says "We've received your request". In which
case, 201 or 202 might be a more appropriate choice.


When autocorrect is enabled, `status: :unprocessable_entity` will be
added to renders used in the `create`, `update`, and `delete` actions.
This is a pretty unreasonable thing to do, but could help people update
these usages en-masse. For this reason, Autocorrect is disabled by
default for this rule.