Set enforced shorthand syntax to always (#78)
Rubocop [changed the
default](https://github.com/rubocop/rubocop/pull/13300) for
`Style/HashSyntax`'s `EnforcedShorthandSyntax` configuration from
`always` to `either`. This reverts that change in `BetterLint` by
explicitly setting the value to `either`.

This is not in line with `Standard`'s config, so I added an exception.

I prefer a consistent enforcement style so we don't have any
pull-request discussion on whether or not to expand/contract hash
syntax. The shorthand syntax also has fewer rooms for typos and makes it
more obvious when there are subtle name differences between hash keys
and values.